requirements.txt
# MultiCord Platform Requirements
# PostgreSQL-centered Discord bot infrastructure with API support

# Core Dependencies
asyncpg>=0.29.0          # PostgreSQL async driver
psutil>=5.9.0            # Process monitoring and health checks
discord.py>=2.3.0        # Discord API integration

# Database & Infrastructure  
asyncio-mqtt>=0.16.0     # For future message queue integration (optional)

# Development & Testing
pytest>=7.4.0           # Testing framework
pytest-asyncio>=0.21.0  # Async test support
black>=23.0.0           # Code formatting
mypy>=1.5.0             # Type checking

# Optional Dependencies for Enhanced Features
# Uncomment as needed for specific deployments

# Template & Configuration Management
# jinja2>=3.1.0          # Template rendering
# pyyaml>=6.0.0          # YAML configuration support

# Enhanced Monitoring
# prometheus-client>=0.17.0  # Metrics export
# structlog>=23.0.0      # Structured logging

# Security & Authentication
cryptography>=41.0.0     # RSA key generation and encryption
pyjwt>=2.8.0            # JWT token creation and validation
argon2-cffi>=23.1.0     # Argon2 password hashing for API keys
python-jose[cryptography]>=3.3.0  # Enhanced JWT handling
passlib[argon2]>=1.7.4  # Password hashing utilities
httpx>=0.25.0           # Async HTTP client for OAuth2 flows

# Web API Framework
fastapi>=0.100.0        # REST API framework
uvicorn[standard]>=0.23.0  # ASGI server with extras
python-multipart>=0.0.6 # Form data parsing
pydantic>=2.0.0         # Data validation and settings
pydantic-settings>=2.0.0  # Settings management

# Additional Security
python-dotenv>=1.0.0    # Environment variable management
keyring>=24.0.0         # Secure credential storage for CLI